Added a Room database to handle data used by the library.
Added a table to store metrics' data and prevent losing data
if the app is killed.

Refactored all the code related to handling metrics to use the
new database instead of the locally cached queue.

Fixed pushUntypedMetric() failing if called before the
bootstrapping process is completed.

Updated javadocs based on new changes in this commit.