Changed the Untyped metric to use a default value of 1. This will help with counting events in Prometheus.
Changed the Wi-Fi session metric to be split into
hourly segments to have more granularity in the data.

Changed the cellular session metric to be split into
hourly segments to have more granularity in the data.

Added a session start field to the Wi-Fi & cellular
session metrics to be able to calculate when did a
session segment ocurr.