Renamed local metric storage attributes to detach from Magma's Prometheus.
Implemented base attributes to include in every
metric that is pushed: Identifier & Timestamp.

Renamed all collected metrics to use camelCase
instead of snake_case.

Modified MetricsWorker to push the metrics to the
new OpenSchema ETL endpoint instead of Magma Orc8r.

Implemented the MetricsPushRequest class to handle
the POST requests to be sent to OpenSchema ETL.