Added DeviceBootReceiver to be able to start the service after a device reboot.
Added ServiceStatusWorker to periodically check
if the service is running correctly or restart
it if necessary.

Added a UUID label to MetricLogsFragment to be
able to debug the data in the cloud. Tapping
the label will copy the UUID to the clipboard.

Added a button to allow the user to manually
trigger a network quality measurement.