refactor(openschema-android): split the Terms and Privacy Policy link
Changed the "Terms and Privacy Policy" link during
onboarding to have 2 separate links "Terms of Service"
and "Privacy Policy" to comply with Play Store
requirements.

Added a new About page to the navigation drawer
which contains the links to both "Terms of Service"
and "Privacy Policy".

Removed the unused FAQ item from the navigation
drawer.